add configuration file for Yun's network configuration
This file can be copied and pasted into the Yun's terminal to configure
it. It sets up a wireless network with the Yun serving as the access
point. Additionally, it forms a bridge connection between the wireless
interface (connection to the computer) and the wired
interface (connection to the Pi)